Add a more expressive compiler interface for resource predicates.
This interface mirrors the scalar predicate compilation interface,
unlike the previous (largely implicit) interface which assumed the
predicate was a `has` predicate and expected the backend to
automatically query the sub-predicate for each field.

This more expressive interface allows resource predicates to have
forms other than `has` (or things that trivially reduce to `has`,
like `is`). For example, the backend now supports an `any`
combinator for resource predicates (although there is not yet any
frontend support for generating such a predicate).

In particular, this will make it possible to add a resource-level
`search` predicate, for #23.